Add possibility to supply additional args to vopt, vlog and vcom
Create a file called vlog_args.txt, vcom_args.txt and/or vopt_args.txt in your
.flycheck-work directory (located in the current directory or
in (projectile-project-root). The text contained within this directory will be
supplied as is to the respective commands. This is very usefull to supply
include directories with +incdir+<path>. If a global (projectile based) work
directory is used (see flycheck-hdl-questasim-use-global-workdir) the relative
path should be relative to the (projectile-project-root). Alternatively absolute
paths can be used as well of course.